operator.h: Add operand priority information in Operator classes
To avoid tests like "if( type == Token::XFX )", we place the
information about the maximum priority an operand can have in
attributes in the classes UnaryOperator and BinaryOperator.

Since these classes are not typedefs anymore, the template Operator
was renamed to OperatorBase.

NullaryOperator was also changed to a class for consistency.