tree_build.cpp: Move NullaryTreeBody parsing to buildExpressionTerminalBody
There was actually a bug in the try-catch chain: if the dynamic_cast
succeed, but the subsequent 'if' fails, we will not try to call
buildExpressionTree recursively to try to convert it either to a
VariableBody or a NumericBody.

Moving the conversion to buildExpressionTerbinalBody solves this
problem and simplifies the code (because every conversion information
for nullary atoms become tied together).