tree_build.cpp: Fix associative operators
The priority retrieval functions in namespace GlobalSymbolTable
return the maximum priority level a subexpression can have.
For the formats FY, YF, XFY and YXF, this returned value
(when refering to the 'Y' side) is exactly the priority of the
operator, so the return value should be compared with '<=', since
the subexpression tree can be rooted by the same operator.

The code was using '<'.