variable.h: Disallow copies and moves of Variable
In theory, it is possible to implement correctly and efficiently
the code for copy, move and assignment of Variable, but since
these operations will not be used (except maybe for cloning),
this extra effort does not worth it.

Note: technically, copy and move constructors and assignment
operators are already "implicitly deleted because the default
definition would be ill-formed".
Explicitly deleting merely improve error messages.