main.cpp: Add operator evaluation to user interface
The semantics of the language says that the last operator should
be nullary and the program evaluation consists in computing this
operator. Computing the last operator is now the default behavior
of the interpreter.

With this commit, we implement all the basic features
of the language in an interpreter.
It remains to implement an interactive mode and fix all bugs.